rename utils to functions in e2e lib
