+++ apps/catalog/catalog-e2e/src/integration/dashboard/movie-tunnel-2.spec.ts
@@ -2,7 +2,7 @@
 
 import { TunnelMainPage, TunnelStorylinePage } from "../../support/pages/dashboard";
 import { signInAndNavigateToMain } from "../../support/utils/utils";
-import { clearDataAndPrepareTest } from "@blockframes/e2e/utils/utils";
+import { clearDataAndPrepareTest } from "@blockframes/e2e/utils/functions";
 
 const CONTENT_TYPES = ['TV Film', 'Library', 'Completed', 'Lagerfeld Confidential', 'Lagerfeld Confidentiel', 'Lagerfeld'];
 const DIRECTORS = ['Rodolphe', 'Marconi', 'Ceci est mon corps (2001)'];
