+++ apps/catalog/catalog-e2e/src/integration/dashboard/movie-tunnel-3.spec.ts
@@ -2,7 +2,7 @@
 
 import { TunnelDistributionRightPage, TunnelStorylinePage, TunnelMainPage } from '../../support/pages/dashboard';
 import { signInAndNavigateToMain } from '../../support/utils/utils';
-import { clearDataAndPrepareTest } from '@blockframes/e2e/utils/utils';
+import { clearDataAndPrepareTest } from '@blockframes/e2e/utils/functions';
 
 // TEST
 const NAVIGATION = ['Title Information', 'Storyline Elements'];
