+++ apps/catalog/catalog-e2e/src/integration/landing/account-creation.spec.ts
@@ -1,7 +1,7 @@
 /// <reference types="cypress" />
 
 import { User } from "@blockframes/e2e/utils/type";
-import { clearDataAndPrepareTest } from "@blockframes/e2e/utils/utils";
+import { clearDataAndPrepareTest } from "@blockframes/e2e/utils/functions";
 import { WelcomeViewPage, LoginViewPage } from "@blockframes/e2e/pages/auth";
 import { OrganizationHomePage } from "@blockframes/e2e/pages/organization";
