+++ apps/catalog/catalog-e2e/src/support/pages/dashboard/TunnelChainOfTitlesPage.ts
@@ -1,5 +1,5 @@
 import TunnelValuationPage from "./TunnelValuationPage";
-import { uploadFile, assertUploadStatus } from "@blockframes/e2e/utils/utils";
+import { uploadFile, assertUploadStatus } from "@blockframes/e2e/utils/functions";
 
 const UPLOAD_STATUS = 'Success';
