+++ apps/catalog/catalog-e2e/src/support/utils/utils.ts
@@ -2,7 +2,7 @@ import { HomePage } from '../pages/marketplace';
 import { TitlesListPage, StartTunnelPage, TunnelMainPage } from '../pages/dashboard';
 import { User } from '@blockframes/e2e/utils/type';
 import { USERS } from '@blockframes/e2e/utils/users';
-import { signIn } from '@blockframes/e2e/utils/utils';
+import { signIn } from '@blockframes/e2e/utils/functions';
 
 // Select user: david.ewing@gillespie-lawrence.fake.cascade8.com
 const LOGIN_CREDENTIALS: Partial<User> = USERS[0];
