+++ libs/e2e/src/lib/utils/functions.ts
