+++ libs/e2e/src/lib/utils/index.ts
@@ -1,4 +1,4 @@
 export * from './type';
 export * from './users';
-export * from './utils';
+export * from './functions';
 export * from './movies';
