+++ tsconfig.json
@@ -52,9 +52,7 @@
       "@blockframes/apps/catalog/marketplace/*": ["apps/catalog/src/app/marketplace/*"],
       "@blockframes/apps/catalog/dashboard/*": ["apps/catalog/src/app/dashboard/*"],
       "@blockframes/apps/catalog/*": ["apps/catalog/catalog/src/app/*"],
-      "@blockframes/event": ["libs/event/src/index.ts"],
       "@blockframes/event/*": ["libs/event/src/lib/*"],
-      "@blockframes/e2e": ["libs/e2e/src/index.ts"],
       "@blockframes/e2e/*": ["libs/e2e/src/lib/*"]
     }
   },
